+++ database/migrations/2019_12_22_174318_create_outputs_table.php
@@ -18,9 +18,9 @@ class CreateOutputsTable extends Migration
             $table->timestamp('date');
             $table->integer('qntd');
 
-            $table->integer('employee_id')->unsigned();
+            $table->bigInteger('employee_id')->unsigned();
             $table->foreign('employee_id')->references('id')->on('employees');
-            $table->integer('product_id')->unsigned();
+            $table->bigInteger('product_id')->unsigned();
             $table->foreign('product_id')->references('id')->on('products');
 
             $table->timestamps();
